pyproject.toml
[project]
name = "vlm-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=11.3.0",
    "torch>=2.7.1",
]
